requirements.txt
pip~=23.2.1
setuptools~=65.5.0
typing_extensions~=4.12.2
SQLAlchemy~=2.0.36
greenlet~=3.1.1
fastapi~=0.115.6
pydantic-settings~=2.7.1
sniffio~=1.3.1
click~=8.1.8
starlette~=0.41.3
pydantic~=2.10.4
pydantic_core~=2.27.2
anyio~=4.7.0
annotated-types~=0.7.0
python-dotenv~=1.0.1
passlib~=1.7.4
cryptography~=44.0.0
Mako~=1.3.8
MarkupSafe~=3.0.2
alembic~=1.14.0
h11~=0.14.0
uvicorn~=0.32.1
PyJWT~=2.10.1
jose~=1.0.0